requirements.txt
pillow~=10.2.0
setuptools~=65.5.1
Cython~=3.0.8
numpy~=1.26.4
pyserial~=3.5
escpos~=1.9
python-escpos[all]~=3.1
python-escpos~=3.1
python-escpos-win~=1.0.1
python-escpos-xml~=0.1.0